feat: cambiar ruta para aceptar jwt

+++ backend/src/main/java/Grupo6/VoluntariadoEmergencias/repositories/AgregarVoluntarioDisponibleRepository.java
@@ -2,5 +2,5 @@ package Grupo6.VoluntariadoEmergencias.repositories;
 
 public interface AgregarVoluntarioDisponibleRepository {
 
-    public void postularComoDisponible(int idEmergencia, String emailVoluntario);
+    public void postularComoDisponible(Long idEmergencia, String emailVoluntario);
 }
